2006-05-10  Paul Eggert  <eggert@cs.ucla.edu>
	* md4.c (rol): Cast right-shift arg to uint32_t to prevent
	unwanted sign propagation, e.g., on hosts with 64-bit int.
	There still are some problems with reeelly weird theoretical hosts
	(e.g., 33-bit int) but it's not worth worrying about now.
	* sha1.c (rol): Likewise.
	(K1, K2, K3, K4): Remove unnecessary L suffix.

+++ lib/md4.c
@@ -280,7 +280,7 @@ md4_process_bytes (const void *buffer, size_t len, struct md4_ctx *ctx)
 #define F(x, y, z) ((z) ^ ((x) & ((y) ^ (z))))
 #define G(x, y, z) (((x) & (y)) | ((x) & (z)) | ((y) & (z)))
 #define H(x, y, z) ((x) ^ (y) ^ (z))
-#define rol(x,n) ( ((x) << (n)) | ((x) >> (32-(n))) )
+#define rol(x, n) (((x) << (n)) | ((uint32_t) (x) >> (32 - (n))))
 #define R1(a,b,c,d,k,s) a=rol(a+F(b,c,d)+x[k],s);
 #define R2(a,b,c,d,k,s) a=rol(a+G(b,c,d)+x[k]+K1,s);
 #define R3(a,b,c,d,k,s) a=rol(a+H(b,c,d)+x[k]+K2,s);
